+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Change Steam Tinker Launch to be installed using the 'steamtinkerlaunch-git' package.
 - Add 'cups-pdf' for modern printer support.
 - Change file system commit time from 30 seconds to 600 seconds.
+- Add support for using the Broadcom proprietary network driver for PCI Ethernet devices and USB adapters.
 
 ## 4.2.0 - The Ventoy Update